pt.block_route_types param: limit connecting modes
Fixes #130

How to verify it's working: on a route that would normally recommend a
bus, add "pt.block_route_types=3" to the request (3 means bus, see GTFS
spec). If there is a train or ferry that serves the trip it should use that
instead. Cafe la Boheme to Erskine Park is an example that works in SF